+++ ui/src/app/core/accessories/types/hap/outlet/outlet.component.ts
@@ -51,4 +51,16 @@ export class OutletComponent {
       void this.service().getCharacteristic!('TargetDoorState').setValue!(this.service().values.TargetDoorState ? 0 : 1)
     }
   }
+
+  public hasCurrentConsumption(): boolean {
+    return 'Consumption' in this.service().values
+  }
+
+  public currentConsumption(): number | undefined {
+    if (!this.hasCurrentConsumption()) {
+      return undefined
+    }
+
+    return this.service().values.Consumption
+  }
 }