+++ ui/src/app/core/accessories/types/hap/valve/valve.component.ts
@@ -120,4 +120,16 @@ export class ValveComponent implements OnInit {
       this.remainingDuration.set('')
     }
   }
+
+  public hasCurrentConsumption(): boolean {
+    return 'Consumption' in this.service().values
+  }
+
+  public currentConsumption(): number | undefined {
+    if (!this.hasCurrentConsumption()) {
+      return undefined
+    }
+
+    return this.service().values.Consumption
+  }
 }